requirements.txt
dj-database-url==0.4.2
Django==2.0.2
django-environ==0.4.4
django-filter==1.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
factory-boy==2.10.0
Faker==0.8.11
Markdown==2.6.11
mysqlclient==1.3.12
PyJWT==1.5.3
pymongo==3.6.0
python-dateutil==2.6.1
python-decouple==3.1
pytz==2018.3
six==1.11.0
text-unidecode==1.1
Unipath==1.1
